requirements.txt
docopt==0.6.2
Flask==3.0.0
ImageHash==4.3.1
Jinja2==3.1.2
more-itertools==10.1.0
Pillow==10.1.0
pymongo==4.6.1
python-magic==0.4.27
termcolor==2.4.0
Werkzeug==3.0.1
Flask-Cors==4.0.0
pybktree==1.1
pillow-heif==0.13.1
av==11.0.0
